Fixed SDL TMRepair broken solution. Manifest and all the dependencies for a plugin were missing.

+++ SDLTMRepair/SDLTMRepair/RepairForm.cs
@@ -10,26 +10,46 @@ namespace Sdl.Community.TMRepair
 {
     public partial class SDLTMRepair : Form
     {
-        public SDLTMRepair()
+	    public readonly string DeployPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), @"SDL Community\TMRepair");
+
+		public SDLTMRepair()
         {
             InitializeComponent();
-        }
+	        var sqlStream = Assembly.GetExecutingAssembly().GetManifestResourceStream("Sdl.Community.TMRepair._3rd_party.sqlite3.exe");
+	        if (sqlStream != null)
+	        {
+		        if (!Directory.Exists(DeployPath))
+		        {
+			        Directory.CreateDirectory(DeployPath);
+		        }
+		        if (!File.Exists(Path.Combine(DeployPath, "sqlite3.exe")))
+		        {
+					var dllPath = Path.Combine(DeployPath, "sqlite3.exe");
+			        var fileStream = File.Create(dllPath, (int)sqlStream.Length);
+			        var assemblyData = new byte[(int)sqlStream.Length];
+			        sqlStream.Read(assemblyData, 0, assemblyData.Length);
+			        fileStream.Write(assemblyData, 0, assemblyData.Length);
+			        fileStream.Close();
+				}  
+	        }
+		}
 
         private void SDLTMRepair_Load(object sender, EventArgs e)
         {
             var waitForm = new WaitForm
             {
-                Width = this.Width - 20,
-                Height = this.Height - 42,
-                Location = new Point(this.Location.X + 10, this.Location.Y + 32),
+                Width = Width - 20,
+                Height = Height - 42,
+                Location = new Point(Location.X + 10, Location.Y + 32),
                 Opacity = 0.7
             };
+
         }
 
         private void btnSelectTM_Click(object sender, EventArgs e)
         {
 	        openFileDialog.Filter = @"Translation memories|*.sdltm";
-            if (openFileDialog.ShowDialog() == DialogResult.OK)
+			if (openFileDialog.ShowDialog() == DialogResult.OK)
             {
                 txtFile.Text = openFileDialog.FileName;
             }
@@ -45,41 +65,45 @@ namespace Sdl.Community.TMRepair
 	        {
 		        txtFile.Enabled = btnIntegrityCheck.Enabled = btnRepair.Enabled = btnSelectTM.Enabled = false;
 
-		        txtLog.Text = "Integrity check started ...";
-		        var path = Environment.GetFolderPath(Environment.SpecialFolder.System);
+		        txtLog.Text = @"Integrity check started ...";
 		        var log = Guid.NewGuid();
 		        var logFolder = Path.GetDirectoryName(txtFile.Text);
-		        var logFilePath = Path.Combine(logFolder, $"log_{log}.txt");
-
-		        var assemblyPath = Assembly.GetExecutingAssembly().Location;
-		        var processStartInfo = new ProcessStartInfo("cmd.exe");
-
-		        processStartInfo.WorkingDirectory = assemblyPath.Substring(0, assemblyPath.LastIndexOf(@"\"));
-		        processStartInfo.Arguments = $@"/c echo .dump | sqlite3.exe ""{txtFile.Text}"" ""pragma integrity_check;"" >""{logFilePath}""";
-		        processStartInfo.UseShellExecute = false;
-
-		        var p = Process.Start(processStartInfo);
-
-		        if (p != null)
+		        if (logFolder != null)
 		        {
-			        while (!p.HasExited)
-			        {
-				        Thread.Sleep(100);
-			        }
+			        var logFilePath = Path.Combine(logFolder, $"log_{log}.txt");
 
-			        var output = string.Empty;
-
-			        using (var sr = new StreamReader(logFilePath))
+			        var processStartInfo = new ProcessStartInfo("cmd.exe")
 			        {
-				        output = sr.ReadToEnd();
+				        WorkingDirectory = DeployPath,
+				        Arguments =
+					        $@"/c echo .dump | sqlite3.exe ""{txtFile.Text}"" ""pragma integrity_check;"" >""{logFilePath}""",
+				        UseShellExecute = false
+			        };
+
+
+			        var p = Process.Start(processStartInfo);
+
+			        if (p != null)
+			        {
+				        while (!p.HasExited)
+				        {
+					        Thread.Sleep(100);
+				        }
+
+				        string output;
+
+				        using (var sr = new StreamReader(logFilePath))
+				        {
+					        output = sr.ReadToEnd();
+				        }
+				        File.Delete(logFilePath);
+				        txtLog.Text += "\r\n";
+				        txtLog.Text += "\r\n";
+				        txtLog.Text += output;
+				        txtLog.Text += "\r\n";
+				        txtLog.Text += "\r\n";
+				        txtLog.Text += @"Integrity check completed.";
 			        }
-			        File.Delete(logFilePath);
-			        txtLog.Text += "\r\n";
-			        txtLog.Text += "\r\n";
-			        txtLog.Text += output;
-			        txtLog.Text += "\r\n";
-			        txtLog.Text += "\r\n";
-			        txtLog.Text += @"Integrity check completed.";
 		        }
 		        txtFile.Enabled = btnIntegrityCheck.Enabled = btnRepair.Enabled = btnSelectTM.Enabled = true;
 			}
@@ -87,7 +111,7 @@ namespace Sdl.Community.TMRepair
 	        {
 				MessageBox.Show(@"Please select an .sdltm file", "", MessageBoxButtons.OK);
 			}
-	     
+	       
         }
 
         private void btnRepair_Click(object sender, EventArgs e)
@@ -101,7 +125,6 @@ namespace Sdl.Community.TMRepair
 		        txtFile.Enabled = btnIntegrityCheck.Enabled = btnRepair.Enabled = btnSelectTM.Enabled = false;
 		        txtLog.Text = @"Repair started ...";
 
-		        var path = Environment.GetFolderPath(Environment.SpecialFolder.System);
 		        var log = Guid.NewGuid();
 		        var sql = Guid.NewGuid();
 
@@ -111,12 +134,15 @@ namespace Sdl.Community.TMRepair
 			        var logFilePath = Path.Combine(logFolder, $"log_{log}.txt");
 			        var sqlFilePath = Path.Combine(logFolder, $"{sql}.sql");
 
-			        var assemblyPath = Assembly.GetExecutingAssembly().Location;
-			        var processStartInfo = new ProcessStartInfo("cmd.exe");
 
-			        processStartInfo.WorkingDirectory = assemblyPath.Substring(0, assemblyPath.LastIndexOf(@"\"));
-			        processStartInfo.Arguments = $@"/c echo .dump | sqlite3.exe ""{txtFile.Text}"" >""{sqlFilePath}""";
-			        processStartInfo.UseShellExecute = false;
+					var processStartInfo = new ProcessStartInfo("cmd.exe")
+			        {
+				        WorkingDirectory = DeployPath,
+
+						Arguments = $@"/c echo .dump | sqlite3.exe ""{txtFile.Text}"" >""{sqlFilePath}""",
+				        UseShellExecute = false
+			        };
+
 
 			        var p = Process.Start(processStartInfo);
 
@@ -127,16 +153,17 @@ namespace Sdl.Community.TMRepair
 					        Thread.Sleep(100);
 				        }
 
-				        var newFileName = Path.Combine(Path.GetDirectoryName(txtFile.Text),
-					        $"repaired_{Path.GetFileName(txtFile.Text)}");
+				        var newFileName = Path.Combine(Path.GetDirectoryName(txtFile.Text), $"repaired_{Path.GetFileName(txtFile.Text)}");
 
-				        var processStartInfo1 = new ProcessStartInfo("cmd.exe");
-				        processStartInfo1.Arguments =
-					        $@"/c sqlite3.exe -init ""{sqlFilePath}"" ""{newFileName}"" "".quit"" >""{logFilePath}""";
+				        var processStartInfo1 = new ProcessStartInfo("cmd.exe")
+				        {
+					        Arguments = $@"/c sqlite3.exe -init ""{sqlFilePath}"" ""{newFileName}"" "".quit"" >""{logFilePath}""",
+					        WorkingDirectory = DeployPath,
+							UseShellExecute = false
+				        };
 
-				        processStartInfo1.WorkingDirectory = assemblyPath.Substring(0, assemblyPath.LastIndexOf(@"\"));
-				        processStartInfo1.UseShellExecute = false;
 				        var p1 = Process.Start(processStartInfo1);
+
 				        if (p1 != null)
 				        {
 					        while (!p1.HasExited)
@@ -145,7 +172,7 @@ namespace Sdl.Community.TMRepair
 					        }
 
 				        }
-				        var output = string.Empty;
+				        string output;
 
 				        using (var sr = new StreamReader(logFilePath))
 				        {
@@ -160,13 +187,12 @@ namespace Sdl.Community.TMRepair
 		        }
 
 		        txtFile.Enabled = btnIntegrityCheck.Enabled = btnRepair.Enabled = btnSelectTM.Enabled = true;
-	        }
+			}
 	        else
 	        {
 		        MessageBox.Show(@"Please select an .sdltm file", "", MessageBoxButtons.OK);
 	        }
-
-           
+			
         }
     }
 }
